Disable CSRF and Sessions in Server
The CSRF token was expected in POSTs to /login, which is unnecessary
given we're not interacting with the server through a browser, and as a
result are not vulnerable to CSRF attacks. This commit disables CSRF
protection.

Likewise, this commit also disables session management. Previously, all
API responses had a `Set-Cookie` header, which is unnecessary since
we're using JWT tokens for authentication.