Remove `linkedStateMixin`
While the mixin gives us some cleaner syntax, it obscures what's
actually happening, so I've removed it. Also, note that even Facebook
states that the mixin "is not needed for most applications and should
be used cautiously."

In addition, there are a handful of minor formatting changes in this
commit.